// src/core/api/generated/spring/schemas/CallRemarkDTOStatus.ts
/**
 * Generated by orval v7.8.0 🍺
 * Do not edit manually.
 * Crm Backend API
 * Crm Backend API documentation
 * OpenAPI spec version: 0.0.1
 */

export type CallRemarkDTOStatus = typeof CallRemarkDTOStatus[keyof typeof CallRemarkDTOStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const CallRemarkDTOStatus = {
  DRAFT: 'DRAFT',
  ACTIVE: 'ACTIVE',
  INACTIVE: 'INACTIVE',
  ARCHIVED: 'ARCHIVED',
} as const;